requirements.txt
pandas
requests
resend
python-dotenv